# gen/api/v1/challenge/challenge.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/challenge/challenge.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/secrets": {
      "get": {
        "summary": "List all secrets in the challenge namespace.",
        "operationId": "SecretManager_ListSecrets",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/challengeListSecretsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "namespace",
            "description": "The namespace to list secrets from. Defaults to the challenge namespace.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "SecretManager"
        ]
      }
    },
...
    "/api/v1/secrets/test-registry": {
      "post": {
        "summary": "Test registry credentials without creating a secret.",
        "operationId": "SecretManager_TestRegistryConnection",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/challengeTestRegistryConnectionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Request to test registry credentials.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/challengeTestRegistryConnectionRequest"
            }
          }
        ],
        "tags": [
          "SecretManager"
        ]
      }
    },
...
    "/api/v1/secrets/{name}": {
      "delete": {
        "summary": "Delete a secret from the challenge namespace.",
        "operationId": "SecretManager_DeleteSecret",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "The secret name.",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "namespace",
            "description": "The namespace where the secret is stored. Defaults to chall-manager.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "SecretManager"
        ]
      }
    }
  },
  "definitions": {
...
    "challengeListSecretsResponse": {
      "type": "object",
      "properties": {
        "secrets": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/challengeSecret"
          },
          "description": "The list of secrets."
        }
      },
      "description": "Response containing the list of secrets."
    },
...
    "challengeSecret": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "gitlab-registry",
          "description": "The secret name."
        },
        "type": {
          "type": "string",
          "example": "kubernetes.io/dockerconfigjson",
          "description": "The secret type (e.g., kubernetes.io/dockerconfigjson, Opaque, kubernetes.io/tls)."
        },
        "namespace": {
          "type": "string",
          "example": "chall-manager",
          "description": "The namespace where the secret is stored."
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "description": "The creation timestamp.",
          "readOnly": true
        },
        "dataKeys": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "The keys stored in the secret.",
          "readOnly": true
        },
        "usedByChallenges": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "List of challenge IDs that use this secret.",
          "readOnly": true
        }
      },
      "description": "A Kubernetes secret representation.",
      "required": [
        "name",
        "type",
        "namespace"
      ]
    },
    "challengeTestRegistryConnectionRequest": {
      "type": "object",
      "properties": {
        "server": {
          "type": "string",
          "example": "registry.gitlab.com",
          "description": "The registry server URL."
        },
        "username": {
          "type": "string",
          "example": "my-username",
          "description": "The registry username."
        },
        "password": {
          "type": "string",
          "example": "glpat-xxx",
          "description": "The registry password or token."
        }
      },
      "description": "Request to test registry credentials.",
      "required": [
        "server",
        "username",
        "password"
      ]
    },
    "challengeTestRegistryConnectionResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "description": "Whether the connection was successful."
        },
        "message": {
          "type": "string",
          "description": "A message describing the result."
        },
        "authenticated": {
          "type": "boolean",
          "description": "Whether authentication was successful."
        }
      },
      "description": "Response from testing registry credentials."
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}